cases/requirements.txt
rlp==1.1.0
dacite==1.0.2
hexbytes==0.1.0
pytest==4.6.6
pytest-forked==0.2
pytest-html==1.19.0
pytest-metadata==1.7.0
pytest-watch==4.2.0
pytest-xdist==1.24.0
requests==2.20.0
client_sdk_python==0.6.9
PyYAML==4.2b4
cryptography==2.4.2
paramiko==2.4.2
allure-pytest==2.8.6
allure-python-commons==2.8.6
ruamel.yaml==0.15.91
pytest-repeat==0.8.0
pytest-rerunfailures==7.0